Add gator serve command: read-only HTTP API for browsing gator state
Dogfooded via gator itself -- a 4-task plan (add-dependencies,
write-serve-tests, implement-handlers, wire-main) dispatched with
`gator dispatch`, all tasks passed with TDD gates.

Endpoints: GET / (HTML), /api/plans, /api/plans/{id}, /api/tasks/{id},
/api/invariants. Uses axum 0.8, tower-http CORS, graceful shutdown.

Also consolidates env-mutex test helpers into a shared test_util module
and fixes COALESCE bigint casts in token usage queries.

+++ crates/gator-db/src/queries/agent_events.rs
@@ -89,8 +89,8 @@ pub async fn list_all_events_for_task(pool: &PgPool, task_id: Uuid) -> Result<Ve
 pub async fn get_token_usage_for_plan(pool: &PgPool, plan_id: Uuid) -> Result<(i64, i64)> {
     let row: (Option<i64>, Option<i64>) = sqlx::query_as(
         "SELECT \
-             COALESCE(SUM((ae.payload->>'input_tokens')::bigint), 0), \
-             COALESCE(SUM((ae.payload->>'output_tokens')::bigint), 0) \
+             COALESCE(SUM((ae.payload->>'input_tokens')::bigint), 0)::bigint, \
+             COALESCE(SUM((ae.payload->>'output_tokens')::bigint), 0)::bigint \
          FROM agent_events ae \
          JOIN tasks t ON t.id = ae.task_id \
          WHERE t.plan_id = $1 AND ae.event_type = 'token_usage'",
@@ -110,8 +110,8 @@ pub async fn get_token_usage_for_plan(pool: &PgPool, plan_id: Uuid) -> Result<(i
 pub async fn get_token_usage_for_task(pool: &PgPool, task_id: Uuid) -> Result<(i64, i64)> {
     let row: (Option<i64>, Option<i64>) = sqlx::query_as(
         "SELECT \
-             COALESCE(SUM((payload->>'input_tokens')::bigint), 0), \
-             COALESCE(SUM((payload->>'output_tokens')::bigint), 0) \
+             COALESCE(SUM((payload->>'input_tokens')::bigint), 0)::bigint, \
+             COALESCE(SUM((payload->>'output_tokens')::bigint), 0)::bigint \
          FROM agent_events \
          WHERE task_id = $1 AND event_type = 'token_usage'",
     )
